Remove Engine trait and rename EngineImpl to Engine
